convert-parquet/requirements.txt
humanize==4.10.0
openpyxl==3.1.4
polars==1.4.1
PyMySQL==1.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.27
